requirements.txt
flask
flask-cors
gunicorn
pillow
scikit-image
-f https://download.pytorch.org/whl/torch_stable.html                                                                                                                                                        
torch==1.5.0+cpu
torchvision==0.6.0+cpu
pdf2image==1.16.3